design : home page - 리스트로 주문 완료
