+++ lib/holders/home/list_order/components/grey_blank.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class GreyBlank extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return SliverToBoxAdapter(
+      child: SizedBox(
+        height: 10,
+        child: Container(
+          color: Colors.grey[100],
+        ),
+      ),
+    );
+  }
+}
